evcc: Enhance API usage bufferSoC
it is possible to subscribe the configuration topic evcc/site/bufferSoC.

This parameter reflects the "priority ev charge" limit from the GUI.
Changes on evcc gui are relfected in that topic, and now being used by evcc-batcontrol.

If charging starts, evcc-plugin sets allow_discharge to this value.
If charging ends, the old value is restored.

To prevent some weird issues with recharging from grid, max_charging_from_grid is lowered (but not restored).

The connection handling is reworked to improve timing issues with retained topics.