evcc: Enhance API usage for battery hold parameters like bufferSoC (#135)
* evcc: Enhance API usage bufferSoC

it is possible to subscribe the configuration topic evcc/site/bufferSoC.

This parameter reflects the "priority ev charge" limit from the GUI.
Changes on evcc gui are relfected in that topic, and now being used by evcc-batcontrol.

If charging starts, evcc-plugin sets allow_discharge to this value.
If charging ends, the old value is restored.

To prevent some weird issues with recharging from grid, max_charging_from_grid is lowered (but not restored).

The connection handling is reworked to improve timing issues with retained topics.

* evcc: align naming

* evcc: store max_charge_limit

In some cases, batcontrol may alter the max_charge_limit lower, if the bufferSoc is below it.

We are storing this value and restore it later.

* evcc: fix max_charge_limit function call

* evcc: fix max_charge_limit comparison logic

* evcc: rename bufferSOC to battery_halt_soc and update related logic

* evcc: max_charge_limit trigger needs to be safed on equal or higher

* evcc: update max_charge_limit logic to store only if old value is higher than new battery_halt_soc