Implementing thread based scheduled refresh
This commit introduces internal scheduling to enable async. data refresh.

Data refresh can happens in the main loop during fetch_forecast or in the scheduled task.
Which comes first, it is holding a lock during refresh.
If a lock is already there, we are skipping the refresh_data part, because the relaxed_caching class should create a possible concurrency.